Devide slices according to functionality
Because it is easy to deal with state, state update
