+++ src/App.test.jsx
@@ -36,9 +36,14 @@ describe('App', () => {
     useDispatch.mockImplementation(() => dispatch);
 
     useSelector.mockImplementation((selector) => selector({
-      prompt,
-      spokenSentence: null,
-      answers: [],
+      application: {
+        answers: [],
+      },
+      speakSentence: {
+        prompt,
+        spokenSentence: null,
+      },
+      yesno: {},
     }));
   });
