+++ src/containers/SetQuestionNumberContainer.jsx
@@ -4,12 +4,12 @@ import { useSelector, useDispatch } from 'react-redux';
 
 import FlowCounter from '../components/FlowCounter';
 
-import { setNumberOfQuestions } from '../redux/slice';
+import { setNumberOfQuestions } from '../redux/slices/applicationSlice';
 
 import { get } from '../utils/utils';
 
 export default function SentenceAnswersContainer() {
-  const numberOfQuestions = useSelector(get('numberOfQuestions'));
+  const { numberOfQuestions } = useSelector(get('application'));
 
   const dispatch = useDispatch();
